Finished User Story 4 - Directions to Classes
User Story 4 is now done.  The user can click on a class from the
Directions Page and get the directions to that class.
I added functions to retrieve the location of the class clicked on,
find the directions to there, and plot them on the map.  Also added
error handling incase the user clicks on an empty class or a class
without a saved location.
Also made changes to the index.html file to incorporate those changes.