Fixes/Additions on the Home Page & Info Pages
Fixed problems with sizing and alignment on the Home Page and Info
pages, plus the Info pages were not linked.
Also added a slot for “Next Class” on the  home page and started
working on the function in dalnav.js to display the student’s next
class (still work in  progress).